질문 삭제 기능 수정(답변 존재 시 예외 처리)

+++ Server/src/main/java/com/stackoverflow/Server/question/service/QuestionService.java
@@ -60,7 +60,12 @@ public class QuestionService {
     public void removeQuestion(long questionId) {
 
         Question question = findQuestion(questionId);
-        questionRepository.deleteById(questionId);
+
+        if (question.getQuestionStatus() == Question.QuestionStatus.QUESTION_NOT_COMMENTED) {
+            questionRepository.deleteById(questionId);
+        } else {
+            throw new BusinessLogicException(ExceptionCode.COMMENT_EXISTS);
+        }
     }
 
     public Page<Question> searchQuestions(String title, Pageable pageable) {
